separate out Dot output to linux only
