+++ linux.config
@@ -1,7 +1,7 @@
 # -*- mode:toml; -*-
 [config]
     # search paths relative to position of the config file
-    moduledirs = [ "mythos/kernel", "prothos"]
+    moduledirs = [ "mythos/kernel", "prothos", "linux"]
     destdir = "linux"
 
     # pseudo modules and so on that are assumed as available
@@ -26,6 +26,7 @@
       #"doxygen",
       #"gitignore",
 	  "runtime-prothos",
+      "linux", 
       "app-init-prothos-linux",
 	  "gcc-linux",
       "app-mlog-prothos-linux",
