+++ linux/app/init.cc
@@ -1,0 +1,1 @@
+../../prothos/platform/linux/app/app/init.cc
