+++ prothos/runtime/runtime/Visitor.hh
@@ -4,8 +4,6 @@
 #include <map>
 #include <vector>
 #include <string>
-#include <sstream>
-#include <fstream>
 
 using namespace Prothos;
 using namespace FlowGraph;
@@ -98,50 +96,4 @@ protected:
 };
 
 
-class DotVisitor: public Visitor {
-public:
-    std::string to_string() {
-        std::stringstream ss;
-        ss << "digraph D {\n";
-        ss << "\n";
-        ss << "rankdir=LR;\n";
 
-        for (auto const& x : visitedNodes) {
-            ss << "\t" << x.second->getKey() << "[shape = " << typeToShape(x.second->getNodeType()) <<"]\n";
-        }
-        ss << "\n";
-        for (auto const& x : visitedNodes) {
-            for (Visitable * v: x.second->getSuccessors()) {
-                // retrieve description for each successor
-                std::map<Visitable*, NodeDescription *>::iterator it = visitedNodes.find(v);
-                NodeDescription * nd = it->second;
-                ss << "\t" << x.second->getKey() << " -> " << nd->getKey() << ";\n";
-            }
-        }
-        ss << "}";
-        return ss.str();
-    }
-
-    void to_file() {
-        std::ofstream myfile;
-        myfile.open ("flowgraph.dot");
-        myfile << to_string();
-        myfile.close();
-    }
-
-    /**
-     * convert nodetype to shapestring
-     **/
-    static std::string typeToShape(NodeType type) {
-        switch(type) {
-            case ENotImplemented: return "star"; break;
-            case EContinueNode: return "box"; break;
-            case EFunctionNode: return "ellipse"; break;
-            case ESplitNode: return "trapezium"; break;
-            case EJoinNode: return "invtrapezium"; break;
-            case ESourceNode: return "doublecircle"; break;
-            case EConditionalNode: return "Mcircle"; break;
-            default: return "circle";
-        }
-    }
-};
